backend/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
asn1crypto==0.24.0        # via cryptography
cffi==1.11.3              # via cryptography
click==6.7
cryptography==2.1.4
envparse==0.2.0
falcon==1.3.0
gunicorn==19.7.1
idna==2.6                 # via cryptography
logzero==1.3.1
pycparser==2.18           # via cffi
pyjwt==1.5.3
python-mimeparse==1.6.0   # via falcon
six==1.11.0               # via cryptography, falcon
ujson==1.35
